Add links to documentation and installation
**Problem**
A few users have asked in #theforeman lately about this plugin, yet they normally arrive at http://projects.theforeman.org/projects/foreman/wiki/Rundeck_Integration and have no guidance about how to proceed since it makes no mention of the plugin.

**Solution**
I've updated the wiki page to point to this repository, and we should consider moving it to a proper manual in theforeman.org. I also tried to update this README with a few more points of contact for help.